pyproject.toml
[project]
name = "zwd-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.47.1",
    "supabase>=2.17.0",
    "openai>=1.57.0",
    "python-dotenv>=1.0.0",
]
